changed float to double

+++ src/main/java/com/hedvig/backoffice/services/autoAnswerSuggestion/AutoAnswerSuggestionServiceStub.java
@@ -47,7 +47,7 @@ public class AutoAnswerSuggestionServiceStub implements AutoAnswerSuggestionServ
     final String intent = "next_fee_date";
     final String reply = "Pengarna dras från ditt bankkonto via autogiro den 27e varje månad";
     final String text = "När dras degen?";
-    final Float confidence = 0.784f;
+    final Double confidence = 0.784;
 
 
     return Lists.newArrayList(new SuggestionDTO(intent, reply, text, confidence, allRepliesArray));
