+++ src/main/java/com/hedvig/backoffice/services/autoAnswerSuggestion/DTOs/SuggestionDTO.java
@@ -10,7 +10,7 @@ public class SuggestionDTO {
   String intent;
   String reply;
   String text;
-  Float confidence;
+  Double confidence;
   ArrayList<ReplyEntryDTO> allReplies;
